+++ tools/apiview/parsers/js-api-parser/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - add `SkipDiff: true` for dependency header line
 - set related line id for pre-release tags
+- add `export const` before constants
 
 # 2.0.2